refactor: rename installation module
